refactor: . move TaskLayoutComponent & taskLayoutComponents to TaskLayoutOptions.ts
