+++ src/Layout/TaskLayoutOptions.ts
@@ -1,4 +1,34 @@
-import { type TaskLayoutComponent, taskLayoutComponents } from './TaskLayout';
+export type TaskLayoutComponent =
+    // NEW_TASK_FIELD_EDIT_REQUIRED
+    | 'description'
+    | 'priority'
+    | 'recurrenceRule'
+    | 'createdDate'
+    | 'startDate'
+    | 'scheduledDate'
+    | 'dueDate'
+    | 'doneDate'
+    | 'cancelledDate'
+    | 'blockedBy'
+    | 'id'
+    | 'blockLink';
+
+// The order here determines the order that task fields are rendered and written to markdown.
+export const taskLayoutComponents: TaskLayoutComponent[] = [
+    // NEW_TASK_FIELD_EDIT_REQUIRED
+    'description',
+    'id',
+    'blockedBy',
+    'priority',
+    'recurrenceRule',
+    'createdDate',
+    'startDate',
+    'scheduledDate',
+    'dueDate',
+    'cancelledDate',
+    'doneDate',
+    'blockLink',
+];
 
 /**
  * Various rendering options of tasks in a query.
