+++ src/Renderer/TaskLineRenderer.ts
@@ -3,11 +3,10 @@ import { Component, MarkdownRenderer } from 'obsidian';
 import { GlobalFilter } from '../Config/GlobalFilter';
 import { TASK_FORMATS, getSettings } from '../Config/Settings';
 import { replaceTaskWithTasks } from '../Obsidian/File';
-import type { TaskLayoutOptions } from '../Layout/TaskLayoutOptions';
+import type { TaskLayoutComponent, TaskLayoutOptions } from '../Layout/TaskLayoutOptions';
 import type { QueryLayoutOptions } from '../Layout/QueryLayoutOptions';
 import type { Task } from '../Task/Task';
 import * as taskModule from '../Task/Task';
-import type { TaskLayoutComponent } from '../Layout/TaskLayout';
 import { StatusMenu } from '../ui/Menus/StatusMenu';
 import { StatusRegistry } from '../Statuses/StatusRegistry';
 import { TaskFieldRenderer } from './TaskFieldRenderer';
