+++ src/TaskSerializer/DataviewTaskSerializer.ts
@@ -1,4 +1,4 @@
-import type { TaskLayoutComponent } from '../Layout/TaskLayout';
+import type { TaskLayoutComponent } from '../Layout/TaskLayoutOptions';
 import type { Task } from '../Task/Task';
 import { Priority } from '../Task/Task';
 import { DefaultTaskSerializer } from './DefaultTaskSerializer';
